requirements.txt
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout<4.0,>=3.0
aiohttp==3.7.4.post0
pandas
discord==2.1.0
attrs==21.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.6
idna==3.2
multidict==5.2.0
praw==7.5.0
prawcore==2.3.0
python-dotenv==0.19.1
requests==2.26.0
tabulate==0.8.9
typing-extensions==3.10.0.2
update-checker==0.18.0
urllib3==1.26.7
websocket-client==1.2.1
yarl==1.7.0